+++ src/scvi/autotune/__init__.py
@@ -1,9 +1,8 @@
 from scvi.utils import error_on_missing_dependencies
 
-error_on_missing_dependencies("hyperopt", "ray.tune", "tensorboard")
+error_on_missing_dependencies("hyperopt", "ray.tune")
 
-
-from ._experiment import AutotuneExperiment, ScibTuneReportCheckpointCallback  # noqa
-from ._tune import run_autotune  # noqa
+from ._experiment import AutotuneExperiment, ScibTuneReportCheckpointCallback  # noqa: E402
+from ._tune import run_autotune  # noqa: E402
 
 __all__ = ["AutotuneExperiment", "ScibTuneReportCheckpointCallback", "run_autotune"]